requirements.txt
flask
streamlit
sentence-transformers
faiss-cpu
PyMuPDF
